api/requirements.txt
flask
mysql-connector-python
oracledb>=1.0.0  # Versión moderna del driver